move state to its own module and file
